Job input is defined in the config.xml file (see the file for its format), and tasks (and their preferred locations) are now correctly parsed out of it.
I refactored the MapReduceJob class to MapReduceJobSpecification.  It is representing how we are specifying input jobs.  It can be constructed directly from the config.xml (just call constructFromXML()).

The JobStatus class has been renamed SimpleJobStatus (and moved to the scheduling.test package) to reflect the fact that it is specific to the FIFOScheduler.  Different schedulers will use different data structures to keep track of jobs.  However, the TaskPool class is still common amongst all schedulers. 

I'm trying to move all extensible framework code into the scheduling package, and all test (user) code in to the scheduling.test package.